requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.0
arxiv==2.0.0
biopython==1.83
Flask-Caching==2.1.0
